DiskManipulator: Clamp partitions to max capacity.
It will not create partitions beyond the available disk space, or partition
table space. The size of the last partition that can be added is clamped to the
available space.

For the `create` command, it now always creates disk images that are the exact
sum of the sizes specified, and the partition sizes will be slightly adjusted as
needed to fit their boot records.

If not all partitions could be created, we inform the user.