Initial implementation of Yamaha SKW-01.
This is the Kanji Word Processing Unit, which has some unique features:
- non-standard kanji font ROM, accessible via non-standard method
- 32kB main program ROM
- 32kB data ROM
- 2kB battery backed SRAM
- printer port, with non-standard I/O mapping

Thanks to several users on MRC, the device could be reverse-engineered in such
a way that it is clear how the happy flow should work. And that was used to
implement the emulation. See the references in the .cc file for the credits.

Thanks to Wouter for reviewing and his suggestions.