MSXtar: Optimise findFirstFreeCluster.
Doing a profile on importing 28 MB worth of files shows that the majority of
time is spent writing data, but the second-most time is spent in
findFirstFreeCluster().

This patch optimises findFirstFreeCluster() to remember and continue from where
it left off, keeping track of the prefix of clusters which it knows are in use,
making the search run in much more linear time rather than exponential.

Before: 39 ms / 510 ms total -> 7.65%
After: 0 ms / 467 ms total -> 0%