Fix crash in SpriteChecker::checkSprites2() when toggling 'VR' bit
Thanks to 'Accumulator' for reporting the issue with an easy to
reproduce test case. See this MRC thread:
    https://www.msx.org/forum/msx-talk/openmsx/openmsx-crashes-feature-to-detect-openmsx

The problem triggered when changing 'sizeMask'. Depending on the 'VR'
bit in VDP R#8, only part of, or all VRAM is accessible (hence more or
less of the 'size' needs to be 'masked').

A key feature of the VDP code is:
* When some property changes, we first render everything up to the
  moment in time right before the change.
* Only then we commit the change.
* (So only future rendering will use the new setting.)

Though for changes in the 'sizeMask' we had a bug: the 'sizeMask'
variable was already changed before we trigger the render callbacks.
But these are now using a mix of old properties and the already updated
'sizeMask'. And this inconsistency caused the crash.

Fixed by only committing the new 'sizeMask' after all (rendering)
callbacks have been triggered.

Likely we never saw this bug before because very few software ever
changes the 'VR' bit.