Fix use-after-free bug in DirectScalerOutput
For example triggered when creating a screenshot with renderer=SDL.
Thanks to Vampier for reporting this bug.

In FBPostProcessor we create a 'stretchScaler' to apply the effect of
the 'horizontal_stretch' setting (via CPU calculations as opposed to GPU
for SDLGL-PP). In the end this writes pixels to a 'SDL_Surface' via the
'SDLDirectPixelAccess' utility. This utility makes sure that we lock and
unlock the SDL_Surface.

We try to reuse this 'stretchScaler' between frames. Because internally
this uses some helper buffers, and we want to avoid having to
allocate/deallocate these buffers on every frame. On some conditions we
do recreate 'stretchScaler': when various settings change, but also when
we render to a different output surface. And the latter happens for
example when we create a screenshot: then we render to a temporary
surface instead of to the display-surface.

Though because of this recycling of 'stretchScaler' we had a bug: we
also kept the 'SDLDirectPixelAccess' for longer than one frame. This had
two consequences:
* We kept the output surface locked for longer than intended. (It should
  be locked/unlocked on every frame).
* When switching between surfaces (e.g. for screenshot), and when
  unlocking the previous surface it could be that this previous surface
  was already deleted. In particular: the temporary surface that was
  used to write the screenshot to is already gone when the next 'normal'
  frame will be rendered.

Fixed by extending the 'ScalerOutput' class with two new methods:
'frameStart()' and 'frameStop()'. In the appropriate subclasses these
methods now respectively allocate/deallocate a 'SDLDirectPixelAccess'
utility object. So this ensures we only keep the surface locked for
(part of) one frame. And we no longer try to unlock an already deleted
surface.