Fix V9990 'incremental' interlaced screen rendering
The V9990 interlaced output was not correct when the screen is rendered
in multiple steps. In some cases the lines do not show up at the correct
vertical position. And with 'multiple steps' I mean: render part of the
screen, then update some register values, and then render more of the
screen.

It took me a while to debug this, but in the end the solution is very
simple. Forgot a multiply by 2 ;-)

The bug comes from a difference between V99x8 and V9990 interlace
rendering. V99x8 flips between two pages, but renders each page
normally. V9990 on the other hand really renders odd and even lines, so
consecutive output lines are 2 lines apart in VRAM.

Thanks to 'Manel Villegas' for reporting this bug. It manifested as
"shaky graphics" when changing the palette (in the middle of a frame).