Fix closing openMSX with ImGui windows outside the main window
OpenMSX reacts to the SDL_QUIT event. SDL sends this event when the
*last* window is closed (e.g. by clicking the 'x' in the corner of the
window). Though when one of the ImGui windows is dragged outside the
main window, there are multiple SDL windows, and then clicking the 'x'
would not close the *last* window, thus no SDL_QUIT event and openMSX
just keeps running (it appears as if clicking the 'x' doesn't have any
effect).

The solution is to also listen to SDL_WINDOWEVENT_CLOSE events. This
event (only) on the main window is now also handled as SDL_QUIT.