Enabled menubar on main openMSX window
For the moment this is a duplicate of the menubar in the main ImGui
window. And using it conflicts with the OSD reversebar, but it's a
proof-of-concept.

Do we want the main ImGui window? Or do we want the main menu bar? Or
maybe configurable? How can this work together with the reversebar?
Maybe rewrite the reverbar as a (dockable?) ImGui widget?