Redo failed MessageCallbacks when fully running.
Some message callbacks are triggered, but cannot be performed when openMSX
isn't fully running yet, because the scripts aren't loaded yet. When a message
callback fails, redo it when openMSX has been fully started.

A little bit hacky, but it solves the problem.

Thanks to Wouter for coming up with the ideas on how to get things connected...

Fixes #1488.