Recreate reverse bar in ImGui
Reimplementation of the OSD reverse bar.
  TODO bookmarks

Part of the motivation was to resolve the conflict between the current
OSD reverse bar and the ImGui main menu bar. This the new version has
extra advantages:
Can be placed at arbitrary location inside or outside the main openMSX
window. Can be docked in other windows.

Right-click to hide the window title and to enable fade-out.

Later this may fully replace the OSD version, but for now kept it to be
able to compare both.
Notice how the ImGui implementation is shorter (simpler?) than the Tcl
version. It's also more efficient and also updated every frame instead
of only 10 times per second.