Added icons window
As a possible replacement for the current OSD icons (Tcl script).

The new icon window can be placed at an arbitrary position, even outside
the main window, or docked in some other window.

Right-click the icon window to configure it. To mimic the old
appearance select "Hide Title".

There's an "Advanced" section where you can fully configure the
appearance. For example disable some icons or select different images.
Right click on a row to reorder, delete or even create new rows.

This new implementation is a lot more flexible than the old Tcl script.
(And it is, within 10 lines, the same number of lines of code).