Fix possible crash when clicking in the imgui reverse bar, part 2
The fix in commit 4679ec3cf00a7d6 didn't go far enough. It postponed the
command till the 'imgui' layer was drawn, but not until all layers were
drawn.

By clicking in the reverse bar (which internally does 'loadstate') we
switch to a new machine. And this includes creating a new VDP-rendering
layer. But that's problematic when we're still iterating over those
layers because we're drawing them.

This patch uses the EventDistributor mechanism to fully postpone all
ImGui-triggered commands till drawing is done.