Reorganize directory structure
"imgui" -> "3rdparty"
Because we may want to add other, non-directly imgui related, libraries
to it.

"src/video/ImGuiLayer.*" -> "src/imgui/*"
Because this file is already large, and will continue to grow. In the
future I'll split it over multiple files. But then it's good to keep
those files grouped in the same directory.