Fix (seamingly) hang on pause/break
When openMSX is paused (or breaked) we don't needlessly re-render the
screen. Though now with the new ImGui stuff not re-rendering also means
that the GUI is irresponsive. And that in combination with ImGui
capturing the events from the rest of openMSX, effectively make it seem
like openMSX hangs.

For now 'fixed' by ensuring a minimal framerate of 10Hz. We may (or may
not) want to revisit this later.