Store our GUI settings in imgui.ini
For example which windows (debuggables, sound settings, ...) were open,
which are the last used paths in file-open-dialogs, or how the OSD icons
or reversebar were configured, are now remembered after closing and
restarting openMSX.

This uses functionality from imgui_internal.h, this API is not
guaranteed to remain the same between different Dear ImGui versions.
Though it is a solution recommended by the Dear ImGui developers
themselves. For example see:
    https://github.com/ocornut/imgui/issues/2564

I'm not happy about the current implementation. The main reason is that
currently we duplicate the to-be-serialized information in two
locations. And we have to keep those in sync. This is especially true
for the more dynamic info (like the 'debuggable' windows that are open,
this list depends on the specific MSX machine). One of the strong points
in the "Immediate Mode GUI paradigm" is that there is no duplicated
information. And thus there cannot be bugs in keeping the information in
sync. The current implementation violates that principle.

The reason why currently the information is duplicated is because of
object lifetime issues. Just before 'Dear ImGui' is shutdown, it still
writes the imgui.ini file, we want our extra information in that file.
But in the current code, when we shutdown 'Dear ImGui' we've already
deleted the 'ImGuiLayer' object which contains the information we want
to save. My first solution was to move the information to a new object
'ImGuiManager' which does have a longer lifetime than 'Dear ImGui'. But
as stated in the previous paragraph: currently this design leads to some
duplicated information that needs to be kept in sync.

I plan to refactor this in the following patches. But I already wanted
to push the current new functionality.

I already wanted to refactor 'ImGuiLayer' so that's it's split over
several smaller files. E.g. separate file for the debugger stuff, the
setting stuff, the media stuff, ... I'll give that higher priority, and
try to do it in a way that also resolves the lifetime issues.