[1/8] Remove SDL renderer
Only keep 'SDLGL-PP' and 'none' renderer.

This patch series only removes the SDL renderer and stuff only used by
that renderer (e.g. base classes that only have one subclass anymore).
This netto removes +45000 lines of code.

For now it keeps all infrastructure to switch renderer in place. This is
_only_ needed to start without renderer (a headless mode), and possibly
activate the renderer later. Maybe in the future we can further simplify
this.

--

Till now we've always kept the old SDL renderer because the code was
written and maintaining it didn't have a (big) cost. That recently
changed: we're experimenting with 'Dear ImGui' to integrate a proper GUI
in openMSX (as a replacement for the OSD menu, catapult and the
debugger). And this needs hardware accelerated graphics, for instance
openGL.