WD2793: tweak detail in writing CRC bytes
DSKPRO-14 sends the following sequence during a 'format track' command:
  F5 F5 F5 FE E2 F7 03 F7
The bytes F5 and F7 have a special meaning:
* F5: writes a 'special' A1 byte, and resets the CRC calculation
* F7: writes two CRC bytes
Thus the above sequence expands into:
  A1 A1 A1 FE E2 6A F5 03 30 63
The first 4 bytes are the sector-header-mark.
The following 4 bytes are the C, H, R, N bytes in the sector header.
And the last two bytes are the CRC-checksum of the header.

Note that the R-field contains the value F5. Normally such sectors
headers cannot be written on a WD2793 becase of the special meaning of
value F5 (normally F5 gets replaced by a 'special' A1).

For more details or background information see:
  https://map.grauw.nl/articles/low-level-disk/

The above is how it's supposed to work, and after this patch openMSX
also emulates it like that. Before there was a bug that F7 (which
outputs the current CRC value) did not correctly update the running CRC
value.