Added ImGuiKeyboard
As an eventual replacement for the "OSD keyboard" (command:
toggle_osd_keyboard).

Activate via 'Settings > Misc > Show virtual keyboard'.

Left-click (and hold) on a virtual keyboard key to press (and hold) the
corresponding MSX key. Releasing left-click also releases the MSX key.
Right-click to keep the key pressed (typically used for 'shift' or
'ctrl').

Most important differences with the OSD keyboard:
* This version can be moved/docked/scaled.
* This version also shows cursors and numerical keypad. In the future
  the layout should become part of the machine/keyboard config.
* This version can be used for input (same as OSD keyboard) but also to
  visualize the current status. E.g. try it in combination with the
  'type' command, or just normal typing on the host keyboard.

Motivation to tackle this now:
A virtual keyboard is not the most important feature for a GUI.
Nevertheless I wanted to implement this now because together with the OSD
menu, the OSD keyboard and menu are two only two users of 'OSDcontrol'
events. In the (near?) future I'd like to simplify the openMSX event
system. Hopefully that'll allow to better integrate ImGui with the
openMSX events (e.g. I'd like to make 'hotkeys' take priority over ImGui
when consuming events). And to make this easier/possible I'd like to
first simplify the openMSX events. In particular drop the '-repeat'
functionality from 'OSDcontrol' events. That should be OK when there are
no active users of 'OSDcontrol' events anymore.