Remove ImGui 'main' window
Move the only remaining item 'ImGui Demo window' from the 'main' ImGui
window to the bottom of 'Settings > Advanced'. That's pretty well
hidden. This demo window is not related to openMSX functionality, but it
is useful while developing new ImGui functionality (it serves as living
documentation about the ImGui functionality).

Now the 'main' ImGui window is empty, except for the menu bar. But the
same menu bar is also present in the main openMSX window. In one of
the previous patches we've disabled the old 'reverse bar', this makes
the menu bar actually useable :-)

Because of the two previous paragraphs, the main ImGui window is no
longer needed. Thus this patch removes it.