Added ImGui based console
I initially tried to have the new console (temporarily) next to the old
console (to be able to better compare them). But having two consoles
didn't interact well with keyboard control. It might have been possible
to fix this, but it wasn't worth the effort.

The new console:
* Can be freely moved, resized and even (un)docked.
* Because of this, settings like 'console{rows,columns,placement}' are
  no longer needed.
* Keyboard handling is now done by 'Dear ImGui' (via a regular InputText
  widget). This means some keybindings are different than before (e.g.
  CTRL-BS instead of ALT-BS to delete the previous word). On the other
  hand, selection and copy-paste within the command line now works as
  expected. Even undo/redo works now (CTRL-Z, CTRL-Y).
* There is no background image anymore in the new console (it also
  wouldn't fit the style of all other ImGui windows). I believe this is
  the only regression in 'functionality' compared to the old console.
* Compared to the old implementation, the new console is only needs
  about half as many lines of code (reduced from ~1500 to ~700).
* The new console likely renders more efficient than the old one: Dear
  ImGui uses a texture-atlas which is initialized once at startup, while
  the old console re-renders (via libfreetype) for every new chunk of
  text.

The interaction with 'HotKeys' is still bad. Currently SDL events are
first passed to 'Dear ImGui'. Only if ImGui doesn't 'consume' the event,
we pass it to the rest of openMSX. OpenMSX uses the old event priority
system. Simplified that is: first HotKeys then the emulated MSX.

This means that, currently, when any ImGui widget is focussed we don't
handle HotKeys. In other words: you can open the console with F10, but
then closing it again with F10 does not work. Instead you first have to
de-focus ImGui (e.g. by clicking in the MSX window), and then F10 works.

I plan to fix this by giving HotKeys event handling higher priority than
ImGui, but that first needs more code refactorings. For now a workaround
could be to undock the console (move it out of the main openMSX window)
and switch between MSX and console via the mouse.