Added debugger tool bar
With buttons to:
* break/continue
* step-in
* step-over
* step-out
* step-back

In the future we might want to extend this with stuff for:
* Symbol manager.
* Breakpoint/watchpoint/condition stuff.
Possibly via a menu bar (instead-of/in addition-to buttons). Possibly
duplicated in the main menu bar.

This patch is also an experiment to add custom icons. Typically in 'Dear
ImGui' icons are drawn as glyphs in a font. To do this it's possible to
add extra custom glyphs in the font. And these can be embedded as a
(compact) array directly in the source code.

We already did this for the icons in the ImGuiFileDialog. What's new in
this patch is that I tried to create the icons myself (inspired by the
corresponding icons in the Qt openMSX debugger).

Roughly the steps are:
* Create .svg files for each of the icons (first time I used 'inkscape').
* Convert the set of .svg files to a .ttf font. To do this I used the
  online conversion tool:
    https://icomoon.io/app/
  Though I'm not 100% happy with this tool, still looking for something
  better. As part of this step you need to assign unicode code-points to
  your icons. The range 0xe000-0xf8ff (Private Use Area (PUA)) is suited
  for this.
* Then convert the .ttf file to a c++ array via the tool:
    https://github.com/ocornut/imgui/blob/master/misc/fonts/binary_to_compressed_c.cpp