Extended 'Settings > Sound' with 'resampler' option
Personally I think this is an advanced setting, and maybe it shouldn't
be exposed in the GUI? At least nowadays, now that most computers are
more than fast enough to handle the 'hq' option.

The main reason I added it is to get the ImGui sound settings menu
on-par with the corresponding menu in the OSD menu. (Or did I still miss
something?)

This patch also extends the EnumSetting-ComboBox widget with a tooltips
map. You can now specify a description for each possible value in the
enum. Currently these descriptions are defined in the GUI-code. Maybe in
the future we may want to move this closer to the setting itself?