Added 'Settings > Video' menu option
This is mostly a recreation of the corresponding stuff in 'catapult' and
the 'OSD menu'. But here are a few notable differences:

* The 'scanline' and 'blur' settings are indented below the
  'scale_algorithm' setting. When the selected algorithm doesn't do
  anything with scanline and/or blur these settings are shown as
  disabled.
* I experimented with the 'TreeNode' Dear ImGui widget. This allows to
  more clearly group settings according to their function (e.g. group
  all settings that have an influence on the colors together). But maybe
  more important, by playing with the default collapsed/expanded state
  of the TreeNodes we can offer 'Advanced' settings without them taking
  up too much space (by default).

This patch also simplifies exposing openMSX settings in ImGui.
Specifying the name of the setting is now optional. By default the name
shown in the GUI is the same as the setting name.