Use I/O range of 4 ports for RTC.
Here is Wouter's analysis to justify this change.
For write, the RP5C01 CS pin is only active for port 0xB5 (and not 0xB4). So
probably this is also the case for read, in other words, it is unlikely that
reading from 0xB4 would be a mirror for reading from 0xB5. Writing to 0xB4
remembers 4 bits (which is typically implemented with a buffer IC). Allowing to
also read these 4 bits back would require extra hardware, and it's not required
by the MSX standard. This is why we assume that reading 0xB4 will have no
effect (will typically return 0xFF, but see below).

I checked whether writing to 0xB4 has any influence on what you read from 0xB4.
And I could not find any effect, on machines with S3527 (NMS 8250, Sony
HB-F6500P) or S1985 (Sony HB-F700D).
I did find out however, that on the latter machine, the output just a loose
MSX-BASIC command: ?HEX$(INP(&HB4)) is different than written in a loop like:
ORA=&HB4TO&HB9:?HEX$(INP(A)):NEXT - see #1355's comments. This is not
understood behaviour and therefore not implemented in the emulation: we just
always return 0xFF, which seems to be correct for some machines, but not for
others.

This makes it easier to implement RTC port mirroring as present in some MSX
ENGINE chips.