Simplify 'Dear ImGui' usage via c++ wrappers
'Dear ImGui' offers a C-API. Many functions need to called in pairs:
e.g. ImGui::Begin() and ImGui::End(), or ImGui::PushID() and
ImGui::PopId(). Not pairing these properly is a bug.

In some cases the 2nd function in such a pair must always be called, in
other cases it may only be called when the first one returned true.
(This inconsistency is documented, and persists only for
backwards compatibility).

This patch simplifies using such paired-functions:
* They will always be properly paired.
* And the inconsistency is hidden.
See comments in src/imgui/ImGuiCpp.hh for more details.

In many cases these function-pairs enable and disable some 'effect'. For
example: ImGui::Indent() and ImGui::Unindent(). With this patch such an
effect gets a scope and then, via the indentation level in the source
code, the scope of the 'effect' is more clearly visible.

Because of the above, best view this patch with 'git diff -b' to ignore
changes in indentation level.