Implement Printer I/O Port mirroring...
...as far as I understand it from #1355.

This means S-3527 based machines have 2 I/O ports, which are mirrored at the
next 6 ports. And S-1985 has 4 I/O ports, which are mirrored at the next 4
ports, and port 0x93 is supposed to be for the bidirectional printing feature,
which was not used in MSX hardware (and thus also not emulated). This
difference is indicated with the new bidirectional parameter (which is default
false, only the S-1985 is known to support it).