Avoid error when clearing joystick mapping config completely.
Setting joystick1_config to {} is not really what you'd want, but let's avoid
showing an error when you open the config in the OSD menu. Perhaps it would be
even better if openMSX would at least make sure that the setting always
contains the dict keys... Let this be a quick workaround for now.