Added ImGui::StrCat() utility
As a (faster) replacement for 'ImGui::Text()' in many (but not all)
cases.

Internally 'ImGui::StrCat()' uses our 'tmpStrCat(...)' function to
effeciently construct a temporary string from sub-parts (these parts can
be strings, integers, ...). Then this temporary string is passed to
'ImGui::TextUnformatted(std::string_view)'.

Some uses of 'ImGui::Text()' remain because it allows for more flexible
format specifiers. In the future we may want to switch to from our
strCat() utility to 'std::format()'. This is both fast (faster than
'sprintf()') and equally flexible (if not more). 'std::format()' is part
of c++20, but for now it needs a too new gcc-version (gcc-13).