Allow symbols in breakpoint viewer
Attempt to use symbol names for breakpoints that were entered on the
openMSX-side.

For breakpoints entered via the GUI-side we keep the
original string formatting, but we now allow to enter symbol names.

Whenever a symbol name is shown, hover over it to see the corresponding
hex value.

TODO what if symbols get removed and a bp was using that symbol
TODO what if a symbol gets a new value (e.g. symbol file reloaded)
TODO what if new symbols get loaded, and a bp was set on such an address