Sync addresses in 'breakpoint viewer' when symbols change
(In this message 'breakpoint' means both 'breakpoint' and 'watchpoint').

When a symbol changes value, breakpoints that use that symbol get
updated.

When a symbol is removed, and there are breakpoints using that symbol,
we replace the address-string in that breakpoint with a hex-address.

When a new symbol is loaded, and the numeric value matches the
hex-address of a breakpoint, we replace that address-string with the new
symbol name. This is a heuristic: we only do this replacement when the
address-string has the exact format "0x....". This covers the cases:
* the bp was entered directly via openMSX commands
* the symbol file was (accidentally) unloaded/reloaded or replaced with
  a different file.
This intentionally does not catch cases were the user manually entered a
matching numeric address but with a different format.