Fix loading old savestates (or replays)
Regression introduced in 6faecb4e47f99.

The 'RomBlocks' class (de)serializes the selected bank (as an offset in
the full ROM) for each of its pages (e.g. a 8kb block mapper has 8
pages). Some of these pages may be unmapped (e.g. a Konami mapper only
maps the region [0x4000,0xc000), but not [0x0000,0x4000) or
[0xc000,0x10000)). Such unmapped pages are represented by the special
offset-value '-1'.

Commit 6faecb4e47f99 changed the type of this offset from 'unsigned' to
'size_t' (see commit message for the motivation). Thus on 64-bit
platforms it changes from 32-bit to 64-bit. But because the type is
unsigned (as in: it has no negative values), the value '-1' is
serialized as '4294967296' (for old savestates) or
'18446744073709551616' (for new savestates). And then when loading an
old savestate, that value '4294967296' is not recognized as the special
value '-1'.

Fixed by also still recognizing '4294967296'. This should be fine
because that value should never be a legal offset value.

I was concerned that savestates created on a 64-bit system would not
load correctly on a 32-bit system (after commit 6faecb4e47f99). But that
doesn't seem to be a problem. Nevertheless we should be more careful in
the future when using 'size_t' in serialized data (or any type whose
size changes between platforms).