[1/10]: remove isRepeatStopper() for OsdControlEvents
The goal of this patch series is to better integrate 'HotKeys' (e.g.
stuff like "bind F9 {toggle fastforward}") with ImGui event handling.
Because ATM HotKeys only work when the openMSX window itself has focus,
not when any ImGui window has focus.

The reason for this behavior is that we currently first send the event
to ImGui and only then to the existing openMSX event infrastructure. We
want to change this to first handle HotKeys, then ImGui and only then
MSX events. (Prior event handlers can block the event for later
handlers).

The challenge with changing this order is that ImGui expects plain SDL
events. But in the current openMSX event infrastructure SDL events have
already been transformed into openMSX specific events (e.g. because we
have additional event types that don't fit well in the SDL event
structure(*1)).
  (*1) Maybe in a future refactoring we could try to use 'SDL_UserEvent'
       though (ATM) I don't think that'll solve all problems.

So in this patch series we'll rework (some of) the openMSX events to
internally retain the original SDL_event. And then we can, half-way
through the openMSX event-pipeline, still pass the original SDL_Event to
ImGui.

--

This patch itself is only a small step towards the above goal. It
removes the "isRepeatStopper()" function for OsdControlEvents (for all
other event types this function already returned always true).

This function was needed to have proper key-repeat behavior for
OsdControlEvents (for "bind -repeat"). Earlier patches in this branch
have already removed the most prominent uses of OsdControlEvents (e.g.
the OSD menu). So having proper key-repeat for them is no longer
required.

The motivation for removing this is that OsdControlEvents are currently
the only events that keep a reference to other events (and this is only
needed in the 'isRepeatStopper()' function). IOW events currently have
"reference semantics", with reference counting, an object pool (with a
lock to be able to create events from different threads). This would all
greatly simplify if we instead use events with "value semantics" (but
then events can't refer to each other anymore).

So in summary: this patch removes functionality that's not really needed
anymore as preparation to simplify the event structure.