[2/10]: remove reference-counting from Event
The commit message in the previous patch already mentioned "reference vs
value semantics" for events. This patch makes the change towards "value
semantics".

Before this patch our 'Event' class was a pointer to a 'RcEvent'
(RC = Reference Counted). And the 'RcEvent' class contained the counter
and the actual event-data, which is a std::variant of all the event
sub-classes (e.g. KeyUpEvent, MouseMotionEvent, ...). After this patch
the 'Event' class is directly that std::variant.

With reference counting, the lifetime of the object it tied to the
reference counter. That is: the object is kept alive for as long as
there is at least one reference to it. Therefor the memory block holding
the object cannot be tied to the lifetime of the reference. One
possibility is to heap-allocate the memory blocks. Another possibility
is to use an ObjectPool. We had chosen for the latter (because we very
frequently allocate/deallocate Event objects). To construct an Event
Object in the ObjectPool we had an Event::create() factory function.
OTOH value-objects are simply stored on the stack or as a member in a
larger struct. Such objects are created via the usual C++ syntax. So a
large part of this patch is to rewrite all uses of 'Event::create()'.

This patch removes the ObjectPool for RcEvent objects. This includes a
mutex that was needed to safely construct Events from different threads.
Now creating (and passing) value-objects in(/between) different threads
doesn't need a mutex anymore.

Copying a reference-counted object only requires copying a pointer and
incrementing the internal counter. Any event type can be copied in this
way. Copying a value object requires making a copy of the full object,
IOW call the objects copy-constructor. At the end of this patch series
(some of) the event types contain an internal SDL_event which is 56
bytes in size. Though, on a modern CPU, copying 56 bytes in one
continuous memory block is likely not more expensive (and possibly even
cheaper) than copying a pointer and incrementing a counter in two
different memory blocks.

For most Event sub classes the copy constructor indeed simply copies a
memory block. For a few it's more involved. In particular the classes
'FileDropEvent' and 'CliCommandEvent' contain a "StringStorage" object
which (currently) cannot be copied, however they can be moved. So part
of this patch now tries to move instead of copy Event objects. In a few
places we did really require a copy, but those places never deal with
'FileDropEvent' or 'CliCommandEvent' objects. So, for now, I gave the
latter two classes a copy constructor which does an "assert(false)". But
if this ever becomes a problem it shouldn't be too hard to fix.

In a few places in the code we needed a null-reference to an Event. That
was supported in the old Event class. But with value-semantics that's no
longer possible. Instead, in the few places that need it, we now use
a 'std::optional<Event>'.

Before we wrote our own 'visit()' utility function to more easily deal
with the std::variant hidden inside 'Event'. Now we directly deal with a
std::variant, so we don't need our own helper anymore. I still kept the
'get()' and 'get_if()' utilities (but renamed them to 'get_event()' and
'get_event_if()' because they do work better with intermediate levels in
the class hierarchy compared to 'std::get()' and 'std::get_if()'.