[3/10]: use SDL2 timestamp in events
A (sub-)goal of this patch series is to replace the internals of our
input events (keyboard, mouse, joystick, ...) with SDL events. One
important difference between our and the SDL events is the timestamp:
* Our timestamp: is 64 bit, in microseconds, obtained from
  std::chrono::steady_clock
* SDL2 timestamp: is 32 bit, in milliseconds, obtained from
  SDL_GetTicks(), depending on the internals of the SDL2 implementation
  this may or may not be the same source as std::chrono::steady_clock.

In openMSX we only actually use this timestamp in a _very_minimal_ way:
It's only used in EventDelay, and effectively only when the 'inputdelay'
settings has a non-zero value (and the default _is_ zero). So the
difference between micro- and milliseconds shouldn't matter (and
milliseconds is still more than accurate enough). (*1)

This patch only replace the 64- with a 32-bit value in the event class.
(Only later patches in this series we'll replace our event internals
with the SDL_Event structure). In the InputEventGenerator class (where
we convert SDL to our events) we now reuse the timestamp already present
in the SDL_Event (instead of querying the current time again). In other
places where we create event objects we either use 0 (when the timestamp
doesn't matter) or use SDL_GetTicks().

So in isolation this patch doesn't do anything useful. It's merely a
preparation to make later patches in this series possible.

--

(*1) In practice the EventDelay class doesn't do anything (anymore).
It's marginally useful when your host machine is too slow to run
openMSX, and you try to work around this by setting a non-zero value for
the 'minframeskip' setting. In such a case a small non-zero 'inputdelay'
value can help to avoid missed key-presses (because the key-press and
corresponding key-release host event get mapped too close together in
emulated time, and then the polling loop in the standard MSX ISR misses
the key). Though this hasn't really been a problem anymore for 10 years
or longer. The only reason to still keep EventDelay is that something
similar may be needed when we want to implement "network-play" for
openMSX (though personally I have no concrete plans to work on this).