[4/10]: switch mouse and joystick input events to SDL_Event
As mentioned in earlier patches: a (sub-)goal of this series is to embed
an SDL_Event structure in our input event objects. This patch makes that
change for mouse and joystick events. (Not yet for keyboard events
because those need a bit more preparation work).

This patch also removes the timestamp from OsdControlEvents because it
wasn't needed. And then (later) only SDL-based events will have a
timestamp.