[5/10]: remove ResizeEvent
This event is sent when the window gets resized. Though currently the
main openMSX window cannot be resized, thus this event never gets sent.

(Only) the SDLVideoSystem class subscribed to this event, but the code
to handle the event was commented out. This is a leftover from an
experiment that DOES make the openMSX window resizable (instead of
having a 'scale_factor' setting with a few fixed values, make the window
freely resizable). I think this is a worthy goal, but it's not something
we'll implement in the short term.

This patch removes the ResizeEvent because it's currently not used (it
does keep the commented-out code in SDLVideoSystem that handles the
event). Moreover one of the next patches in this series will expose the
SDL_WindowEvent, that one includes window-resizing. So in the future
SDLVideoSystem can switch to that.