[9/10]: added TextEvent
This was the last missing SDL2 event that's relevant for ImGui.

Before both key- and text-events were combined in a single openMSX
event. We (still) have an heuristic to include unicode information
(coming from the text-event) into the key-event. But now, in addition,
openMSX also has a text-event.

For example pressing the 'A' key produces both a key-event and a
text-event. The hotkey subsystem currently listens for key-events. If
there was something bound to the 'A' key, then this event is 'consumed'.
Meaning it's no longer send to event-listeners with lower priority.
Before this patch this would 'fully' block the event associated with
this 'A' key. But now we only block the key-event, and not the
text-event.

For openMSX itself the above doesn't matter: openMSX itself doesn't
listen to text-events. But it does matter for ImGui. In particular
because the last patch in this series will give the HotKey event-
listener higher priority than the ImGui event-listener. So that means
text-input in ImGui widgets can no longer be blocked because of bound
hotkeys. Initially I thought this was an unwanted side-effect. But now
that I've tried it out, I think it might actually be a good thing. We'll
see.