[10/10]: give HotKey higher priority than ImGui
This was the main goal of this patch series (the previous patches were
only preparations to make this possible).

The 'Dear ImGui' library can integrate with various windowing systems,
including SDL2. In this integration you have to pass SDL2 events to
ImGui. That way ImGui can react to mouse clicks, keyboard input, window
movements, ...

Up till now we did exactly that: in the openMSX eventloop, at the place
where we poll for SDL2 events, we pass this event to ImGui and then
continue with the rest of the (existing) openMSX event handling code.
The big problem with this approach was that ImGui can (indicate that it
wants to) consume this event. And then the event shouldn't be further
processed by the rest of the application. Usually that's good: e.g. when
you're entering text in an ImGui edit field, you don't also want those
key pressed to go to the emulated MSX. Though there are also situations
where this doesn't work as well: for example when some ImGui window has
focus, this also blocks the events and then stuff like HotKeys (e.g. F12
to mute) don't work anymore. before this patch you first had to give the
main openMSX window focus before hotkeys would work. That's annoying.

In this patch we can give the HotKey subsystem higher priority than
ImGui. So ImGui can no longer block events for HotKey, but it can still
block events from the MSX emulation. To achieve this we had to rework
how events are represented in openMSX. Before we transformed SDL2 events
to another structure that's better suited for openMSX (e.g. because we
have extra event types and/or we want automatic memory management for
some of those (extra) event types). After this patch we still have our
own event types, but (some of them) are thin wrappers around SDL2
events. Meaning we can still get the original SDL2 event out of the
openMSX event. And that allowed to integrate ImGui in the _middle_ of
our event-handling-pipeline (instead of at the front).